Giant refactor and cleanup. (#83)
* Giant refactor and cleanup.

* Add invariances

* Fix tests.
