+++ timemachine/potentials/jax_utils.py
